+++ src/benchmark/frameworks/openvino/openvino_benchmark_process.py
@@ -1,0 +1,54 @@
+import re
+
+from benchmark.frameworks.openvino.openvino_process import OpenVINOProcess
+
+
+class OpenVINOBenchmarkProcess(OpenVINOProcess):
+    def __init__(self, test, executor, log, perf_hint=''):
+        super().__init__(test, executor, log)
+        self._perf_hint = perf_hint
+
+    @staticmethod
+    def _add_perf_hint_for_cmd_line(command_line, perf_hint):
+        hint = perf_hint.lower()
+        if hint in ('latency', 'throughput'):
+            return f'{command_line} -hint {hint}'
+        return command_line
+
+    @staticmethod
+    def _add_extension_for_cmd_line(command_line, extension, device):
+        if 'GPU' in device:
+            return f'{command_line} -c {extension}'
+        elif 'CPU' in device or 'MYRIAD' in device:
+            return f'{command_line} -l {extension}'
+        return command_line
+
+    def get_performance_metrics(self):
+        if self._status != 0 or len(self._output) == 0:
+            return None, None, None
+
+        # calculate average time of single pass metric to align output with custom launchers
+        duration = self._get_benchmark_app_metric('Duration')
+        iter_count = self._get_benchmark_app_metric('Count')
+        average_time_of_single_pass = (round(duration / 1000 / iter_count, 3)
+                                       if None not in (duration, iter_count) else None)
+
+        fps = self._get_benchmark_app_metric('Throughput')
+        latency = round(self._get_benchmark_app_metric('Median') / 1000, 3)
+
+        return average_time_of_single_pass, fps, latency
+
+    def _get_benchmark_app_metric(self, metric_name):
+        """
+        gets metric value from benchmark app full output
+        :param metric_name: metric name, ex 'Throughput'
+        :return: float value or None if pattern not found
+        """
+        for line in self._output:
+            regex = re.compile(f'.*{metric_name}:\\s+(?P<metric>\\d*\\.\\d+|\\d+).*')
+            res = regex.match(line)
+            if res:
+                try:
+                    return float(res.group('metric'))
+                except ValueError:
+                    return None
