+++ src/benchmark/frameworks/openvino/openvino_process.py
@@ -1,0 +1,12 @@
+from abc import ABC
+
+from benchmark.processes import ProcessHandler
+
+
+class OpenVINOProcess(ProcessHandler, ABC):
+    def __init__(self, test, executor, log):
+        super().__init__(test, executor, log)
+
+    @staticmethod
+    def __add_nthreads_for_cmd_line(command_line, nthreads):
+        return f'{command_line} -nthreads {nthreads}'
